refactor: move functions to utility file and clean up calling function
