+++ src/app/Layout/Presentation/PresentationComponent.tsx
@@ -1,13 +1,17 @@
 ﻿import * as React from "react";
 import { PresentationType } from "../Types/PresentationType";
-import { LagueDhontResult, PartyResult, DistrictResult, PartyRestQuotients, LevelingSeat } from "../Interfaces/Results";
+import { LagueDhontResult, PartyResult, DistrictResult } from "../Interfaces/Results";
 import { ElectionOverview, DistrictOverview, SeatsPerParty, SeatDistribution, SingleDistrict } from "./Views";
 import {
     getDistrictTableData,
     getPartyTableData,
     getSeatDistributionData,
     getSeatsPerPartyData,
-    roundPartyResults
+    roundPartyResults,
+    flattenPartyRestQuotients,
+    removeSeatDuplicates,
+    sortSeatsByNumber,
+    getRoundsAssignedSeats
 } from "./Utilities/PresentationUtilities";
 import { RemainderQuotients } from "./Views/RemainderQuotients";
 
@@ -73,54 +77,10 @@ export class PresentationComponent extends React.Component<PresentationProps, {}
     }
 
     getLevellingSeats() {
-        /**
-         * Flattens a cumbersome object
-         * @param prqs a container object that needs to be flattened
-         */
-        function flattenPartyRestQuotients(prqs: PartyRestQuotients[]): LevelingSeat[] {
-            const levellingSeats: LevelingSeat[] = [];
-            prqs.forEach((prq) => {
-                prq.levelingSeats.forEach((seat) => {
-                    levellingSeats.push(seat);
-                });
-            });
-            return levellingSeats;
-        }
-        const seats: LevelingSeat[] = [];
-        const seatRounds = flattenPartyRestQuotients(this.props.results.levelingSeatDistribution);
-        console.log("seatRounds", seatRounds);
-        seatRounds.forEach((round) => {
-            if (round.seatNumber > 0) {
-                seats.push(round);
-            }
-        });
-        function removeSeatDuplicates(seats: LevelingSeat[]): LevelingSeat[] {
-            const existingSeatsSet: Set<string> = new Set();
-            const uniqueSeats: LevelingSeat[] = [];
-            seats.forEach((seat) => {
-                if (!existingSeatsSet.has(seat.district + seat.seatNumber)) {
-                    uniqueSeats.push(seat);
-                    existingSeatsSet.add(seat.district + seat.seatNumber);
-                }
-            });
-            return uniqueSeats;
-        }
-
-        function sortSeatsByNumber(seats: LevelingSeat[]): LevelingSeat[] {
-            const sortedSeats: LevelingSeat[] = seats;
-            sortedSeats.sort((a, b) => {
-                if (a.seatNumber < b.seatNumber) {
-                    return -1;
-                } else if (a.seatNumber > b.seatNumber) {
-                    return 1;
-                } else {
-                    return 0;
-                }
-            });
-            return sortedSeats;
-        }
-        console.log("levelling seats:", sortSeatsByNumber(removeSeatDuplicates(seats)));
-        return sortSeatsByNumber(removeSeatDuplicates(seats));
+        const flattened = flattenPartyRestQuotients(this.props.results.levelingSeatDistribution);
+        const assignedSeats = getRoundsAssignedSeats(flattened);
+        const noDuplicateSeats = removeSeatDuplicates(assignedSeats);
+        return sortSeatsByNumber(noDuplicateSeats);
     }
 
     render() {
